rename :  에러바운더리 관련 shared/ui이전
