--- client/src/features/boundary/index.ts
@@ -1,1 +1,0 @@
-export { default as ErrorBoundary } from './ErrorBoundary';
