+++ client/src/shared/ui/error-boundary/ErrorBoundary.tsx
