+++ client/src/shared/ui/error-boundary/error-code.ts
