+++ client/src/shared/ui/error-boundary/index.ts
@@ -1,0 +1,2 @@
+export { default as ErrorBoundary } from './ErrorBoundary';
+export { default as GlobalErrorBoundary } from './GlobalErrorBoundary';
