+++ client/src/shared/ui/error-boundary/ui/ConnectionError.tsx
