+++ client/src/shared/ui/error-boundary/ui/RequireLogin.tsx
