+++ client/src/shared/ui/error-boundary/ui/styled.ts
